[common] Move pkg/common/config -> pkg/config
