+++ pkg/agent/config/server.go
@@ -2,7 +2,7 @@ package config
 
 import (
 	iconfig "github.com/at15/go.ice/ice/config"
-	cconfig "github.com/benchhub/benchhub/pkg/common/config"
+	cconfig "github.com/benchhub/benchhub/pkg/config"
 )
 
 type CentralConfig struct {
